fixed issue with ActivityUserProfile intent tests
had to split the intent tests into 2 classes
